# fhir-iq-prior-auth/specs/api/openapi.yaml
openapi: 3.0.3
info:
  title: FHIR IQ Prior Authorization Support API
  description: |
    FHIR-based Prior Authorization Support (PAS) API implementing Da Vinci PAS and DTR
    Implementation Guides for CMS-0057-F compliance demonstration.

    This API provides endpoints for:
    - Submitting prior authorization requests (PAS)
    - Inquiring about authorization status
    - Documentation Templates and Rules (DTR) support
    - Provider and Patient Access APIs
    - Payer-to-payer bulk export operations
  version: 1.0.0-poc
  contact:
    name: FHIR IQ Integration Team
    email: integration@fhir-iq.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: https://pas-api.fhir-iq.com/fhir
    description: Production FHIR IQ PAS Server
  - url: http://localhost:3000/fhir
    description: Local development server
security:
  - OAuth2:
      - system/Claim.write
      - user/Claim.read
      - patient/Claim.read
paths:
  /Claim/$submit:
    post:
      summary: Submit Prior Authorization Request
      description: |
        Submit a prior authorization request Bundle containing a Claim resource
        and supporting documentation per Da Vinci PAS IG requirements.
      operationId: submitPriorAuth
      security:
        - OAuth2:
            - system/Claim.write
      requestBody:
        required: true
        content:
          application/fhir+json:
            schema:
              $ref: '#/components/schemas/Bundle'
            example:
              resourceType: Bundle
              type: collection
              entry:
                - resource:
                    resourceType: Claim
                    status: active
                    type:
                      coding:
                        - system: http://terminology.hl7.org/CodeSystem/claim-type
                          code: professional
      responses:
        '200':
          description: Synchronous authorization response
          content:
            application/fhir+json:
              schema:
                $ref: '#/components/schemas/Bundle'
        '202':
          description: Asynchronous processing initiated
          content:
            application/fhir+json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: Invalid request or validation errors
          content:
            application/fhir+json:
              schema:
                $ref: '#/components/schemas/OperationOutcome'
        '401':
          description: Authentication required
          content:
            application/fhir+json:
              schema:
                $ref: '#/components/schemas/OperationOutcome'
        '403':
          description: Insufficient permissions
          content:
            application/fhir+json:
              schema:
                $ref: '#/components/schemas/OperationOutcome'
  /Claim/$inquire:
    post:
      summary: Inquire Prior Authorization Status
      description: |
        Check the status of a previously submitted prior authorization request
        using claim identifier, patient reference, or authorization number.
      operationId: inquirePriorAuth
      security:
        - OAuth2:
            - user/Claim.read
            - system/Claim.read
      requestBody:
        required: true
        content:
          application/fhir+json:
            schema:
              type: object
              properties:
                resourceType:
                  type: string
                  enum: [Parameters]
                parameter:
                  type: array
                  items:
                    type: object
                    properties:
                      name:
                        type: string
                        enum: [identifier, patient, provider, authorizationNumber]
                      valueIdentifier:
                        $ref: '#/components/schemas/Identifier'
                      valueReference:
                        $ref: '#/components/schemas/Reference'
                      valueString:
                        type: string
      responses:
        '200':
          description: Authorization status found
          content:
            application/fhir+json:
              schema:
                $ref: '#/components/schemas/Bundle'
        '404':
          description: Authorization not found
          content:
            application/fhir+json:
              schema:
                $ref: '#/components/schemas/OperationOutcome'
...
components:
  securitySchemes:
    OAuth2:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: https://auth.fhir-iq.com/oauth2/authorize
          tokenUrl: https://auth.fhir-iq.com/oauth2/token
          scopes:
            system/Claim.read: Read claim resources (system)
            system/Claim.write: Write claim resources (system)
            user/Claim.read: Read claim resources (user)
            user/ClaimResponse.read: Read claim response resources (user)
            patient/Claim.read: Read patient's claim resources
            patient/ClaimResponse.read: Read patient's claim responses
            user/Questionnaire.read: Read questionnaire resources
            user/QuestionnaireResponse.write: Create questionnaire responses
            system/Patient.read: Read patient resources (system)
            system/export: Perform bulk export operations
        clientCredentials:
          tokenUrl: https://auth.fhir-iq.com/oauth2/token
          scopes:
            system/Claim.write: Write claim resources (system)
            system/export: Perform bulk export operations
  schemas:
    Bundle:
      type: object
      properties:
        resourceType:
          type: string
          enum: [Bundle]
        type:
          type: string
          enum: [document, collection, searchset, batch-response]
        total:
          type: integer
        entry:
          type: array
          items:
            type: object
    Task:
      type: object
      properties:
        resourceType:
          type: string
          enum: [Task]
        status:
          type: string
          enum: [draft, requested, received, accepted, in-progress, completed, failed]
        focus:
          $ref: '#/components/schemas/Reference'
...
    OperationOutcome:
      type: object
      properties:
        resourceType:
          type: string
          enum: [OperationOutcome]
        issue:
          type: array
          items:
            type: object
            properties:
              severity:
                type: string
                enum: [fatal, error, warning, information]
              code:
                type: string
              details:
                $ref: '#/components/schemas/CodeableConcept'
    Identifier:
      type: object
      properties:
        use:
          type: string
          enum: [usual, official, temp, secondary]
        system:
          type: string
          format: uri
        value:
          type: string
    Reference:
      type: object
      properties:
        reference:
          type: string
        display:
          type: string
    CodeableConcept:
      type: object
      properties:
        coding:
          type: array
          items:
            type: object
            properties:
              system:
                type: string
                format: uri
              code:
                type: string
              display:
                type: string
        text:
          type: string
...